@@ -35,8 +35,8 @@ export class ReportsComponent implements OnInit {
     categories: Category[] = [];
     entries: Entry[] = [];
 
-    @ViewChild('month') month: ElementRef = null;
-    @ViewChild('year') year: ElementRef = null;
+    @ViewChild('month', { static: true }) month: ElementRef = null;
+    @ViewChild('year', { static: true }) year: ElementRef = null;
 
     constructor(private entryService: EntryService, private categoryService: CategoryService) {
     }
